Auth/iflogged

+++ client/src/pages/DeleteCardSet.jsx
@@ -1,26 +1,34 @@
 import { useMutation } from '@apollo/client';
 import { DELETE_CARDSET } from '../utils/mutations';
-
+import Auth from '../utils/auth';
+// delete will be a button on the card set page
+// export the button to the card set page
+// we will not use this as a whole page
 
 const DeleteCardSet = ({ id }) => {
     const [deleteCardSet, { error }] = useMutation(DELETE_CARDSET);
-  
     const handleDelete = async () => {
+      
       try {
-        const { data } = await deleteCardSet({
-          variables: {
-            id: id
-          }
-        });
+        const confirmDelete = window.confirm('Deleting this card is permanent. Are you sure you want to delete it?');
+        if (confirmDelete) {
+            const { data } = await deleteCardSet({
+                variables: {
+                    id: id
+                }
+            });    
+
+        };
         console.log('Flashcard set deleted:', data.deleteCardSet);
       } catch (error) {
         console.error('Error deleting flashcard set:', error);
       }
     };
 
+    if (Auth.loggedIn()) {
     return (
       <button onClick={handleDelete}>Delete Card Set</button>
     );
   };
-
+};
   export default DeleteCardSet;
